Change noise on development rates
Changed the way that noise was added so that A(t+1) = A(t) + LN(dA*dt)
as opposed to A(t+1) = LN(A(t) + da*dt). This ensures that individuals
will never regress in development. Additionally, altered the way kernels
were calculated to be cleaner and easier to understand (the biggest
impact is on the diapause kernel).